Reorganize descriptors into their own files
Fix a bunch of incorrect docstrings.

Add missing attributes to the MeshDescriptor class that are
expected for all subclasses.

Add missing installation requirements to setup.py
